backend/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.5.1",
    "email-validator>=2.0",
    "pyjwt>=2.10.1",
    "psycopg2-binary==2.9.10",
    "python-dotenv>=1.2.1",
    "python-dateutil>=2.9.0.post0",
    "redis>=5.0.0",
    "uvicorn>=0.40.0",
    "factory-boy>=3.3.3",
    "faker>=40.4.0",
]

[dependency-groups]
dev = [
    "pytest >=8.1.1,<9",
    "pytest-django>=4.11.1",
    "ruff>=0.14.10",
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py313"
exclude = ["migrations"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]
ignore = []

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint.isort]
known-first-party = ["api"]
force-single-line = false

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'config.test_settings'
addopts = '--reuse-db'
